+++ app/controllers/feedback_files_controller.rb
@@ -4,7 +4,7 @@ class FeedbackFilesController < ApplicationController
 
   before_action { authorize! }
 
-  def get_feedback_file
+  def show
     feedback_file = FeedbackFile.find(params[:id])
     send_data_download feedback_file.file_content, filename: feedback_file.filename
   end