+++ app/models/feedback_file.rb
@@ -1,6 +1,8 @@
 class FeedbackFile < ApplicationRecord
   belongs_to :submission, optional: true
-  belongs_to :test_run, optional: true
+  belongs_to :test_group_result, optional: true
+
+  validates_presence_of :submission_id, if: -> { test_group_result_id.nil? }
 
   validates_presence_of :filename # we need a filename
   validates_presence_of :mime_type # we need a mime type
@@ -14,5 +16,14 @@ class FeedbackFile < ApplicationRecord
     return false if new_content.nil?
     self.file_content = new_content
     self.save
+  end
+
+  # Returns the associated grouping for this feedback file
+  def grouping
+    if submission_id.nil?
+      test_group_result.test_run.grouping
+    else
+      submission.grouping
+    end
   end
 end