+++ db/migrate/20210505032404_add_test_group_result_id_to_feedback_files.rb
@@ -1,0 +1,5 @@
+class AddTestGroupResultIdToFeedbackFiles < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :feedback_files, :test_group_result, index: true, foreign_key: true
+  end
+end