+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_04_27_120039) do
+ActiveRecord::Schema.define(version: 2021_05_05_032404) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -203,9 +203,9 @@ ActiveRecord::Schema.define(version: 2021_04_27_120039) do
     t.integer "submission_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.bigint "test_run_id"
+    t.bigint "test_group_result_id"
     t.index ["submission_id"], name: "index_feedback_files_on_submission_id"
-    t.index ["test_run_id"], name: "index_feedback_files_on_test_run_id"
+    t.index ["test_group_result_id"], name: "index_feedback_files_on_test_group_result_id"
   end
 
   create_table "grace_period_deductions", id: :serial, force: :cascade do |t|
@@ -617,7 +617,7 @@ ActiveRecord::Schema.define(version: 2021_04_27_120039) do
   add_foreign_key "extensions", "groupings"
   add_foreign_key "extra_marks", "results", name: "fk_extra_marks_results", on_delete: :cascade
   add_foreign_key "feedback_files", "submissions"
-  add_foreign_key "feedback_files", "test_runs"
+  add_foreign_key "feedback_files", "test_group_results"
   add_foreign_key "grader_permissions", "users"
   add_foreign_key "grouping_starter_file_entries", "groupings"
   add_foreign_key "grouping_starter_file_entries", "starter_file_entries"