+++ spec/factories/feedback_files.rb
@@ -1,5 +1,3 @@
-require 'faker'
-
 FactoryBot.define do
   factory :feedback_file do
     association :submission
@@ -9,7 +7,7 @@ FactoryBot.define do
   end
 
   factory :feedback_file_with_test_run, class: FeedbackFile do
-    association :test_run, factory: :student_test_run
+    association :test_group_result
     filename { "#{Faker::Lorem.word}.txt" }
     mime_type { 'text/plain' }
     file_content { Faker::Lorem.sentence }